change: rename `network-v1` crate to `legacy`
Rename the `openraft-network-v1` crate to `openraft-legacy` to serve as a
general compatibility layer for deprecated Openraft APIs. The V1 network API
is now organized in the `network_v1` submodule, allowing future legacy APIs
to be added as separate modules.

Changes:
- Rename directory `network-v1/` to `legacy/`
- Rename package from `openraft-network-v1` to `openraft-legacy`
- Restructure source files into `src/network_v1/` submodule
- Rename `adapt_v1_to_v2.rs` to `adapter.rs` for clarity
- Update all dependent crates to use `openraft-legacy`
- Update imports from `openraft_network_v1` to `openraft_legacy::network_v1`
- Update documentation references in openraft core

Upgrade tip:

Update your `Cargo.toml` dependency:
- `openraft-network-v1 = "0.10"` → `openraft-legacy = "0.10"`

Update your imports:
- `use openraft_network_v1::RaftNetwork` → `use openraft_legacy::network_v1::RaftNetwork`
- `use openraft_network_v1::Adapter` → `use openraft_legacy::network_v1::Adapter`
- `use openraft_network_v1::ChunkedRaft` → `use openraft_legacy::network_v1::ChunkedRaft`
