+++ benchmarks/minimal/Cargo.toml
@@ -18,7 +18,7 @@ repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
 openraft            = { path = "../../openraft", version = "0.10.0", features = ["serde", "type-alias", "runtime-stats"] }
-openraft-network-v1 = { path = "../../network-v1", version = "0.10.0" }
+openraft-legacy = { path = "../../legacy", version = "0.10.0" }
 
 anyhow     = { version = "1.0.63" }
 clap       = { version = "4.2", features = ["derive"] }
