+++ examples/network-v1-http/src/lib.rs
@@ -17,8 +17,8 @@ use openraft::raft::InstallSnapshotRequest;
 use openraft::raft::InstallSnapshotResponse;
 use openraft::raft::VoteRequest;
 use openraft::raft::VoteResponse;
-use openraft_network_v1::Adapter;
-use openraft_network_v1::RaftNetwork;
+use openraft_legacy::network_v1::Adapter;
+use openraft_legacy::network_v1::RaftNetwork;
 use reqwest::Client;
 use serde::Serialize;
 use serde::de::DeserializeOwned;
