+++ examples/raft-kv-memstore-single-threaded/Cargo.toml
@@ -17,7 +17,7 @@ repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
 openraft           = { path = "../../openraft", features = ["serde", "single-threaded", "type-alias"] }
-openraft-network-v1 = { path = "../../network-v1" }
+openraft-legacy = { path = "../../legacy" }
 
 futures            = { version = "0.3" }
 serde              = { version = "1.0.114", features = ["derive"] }
