+++ examples/raft-kv-memstore-single-threaded/src/app.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 use futures::StreamExt;
 use futures::channel::mpsc;
 use futures::channel::oneshot;
-use openraft_network_v1::ChunkedRaft;
+use openraft_legacy::network_v1::ChunkedRaft;
 
 use crate::NodeId;
 use crate::StateMachineStore;
