+++ examples/raft-kv-rocksdb/Cargo.toml
@@ -24,7 +24,7 @@ path = "src/bin/main.rs"
 client-http         = { path = "../client-http" }
 network-v1-http     = { path = "../network-v1-http" }
 openraft            = { path = "../../openraft", features = ["serde", "type-alias"] }
-openraft-network-v1 = { path = "../../network-v1" }
+openraft-legacy = { path = "../../legacy" }
 openraft-rocksstore = { path = "../rocksstore" }
 types-kv            = { path = "../types-kv" }
