+++ examples/raft-kv-rocksdb/src/lib.rs
@@ -9,7 +9,7 @@ use actix_web::middleware;
 use actix_web::middleware::Logger;
 use actix_web::web::Data;
 use openraft::Config;
-use openraft_network_v1::ChunkedRaft;
+use openraft_legacy::network_v1::ChunkedRaft;
 
 use crate::app::App;
 use crate::network::api;
