+++ legacy/src/network_v1/mod.rs
@@ -1,6 +1,6 @@
-//! RaftNetwork V1 adapter for Openraft with chunked snapshot support.
+//! Network V1 adapter for Openraft with chunked snapshot support.
 //!
-//! This crate provides backward compatibility for applications using the v1
+//! This module provides backward compatibility for applications using the v1
 //! `RaftNetwork` trait with chunk-based snapshot transmission.
 //!
 //! # Components
@@ -10,7 +10,7 @@
 //! [`Adapter`] wraps any `RaftNetwork` (v1) implementation to provide `RaftNetworkV2`.
 //!
 //! ```ignore
-//! use openraft_network_v1::{Adapter, RaftNetwork};
+//! use openraft_legacy::network_v1::{Adapter, RaftNetwork};
 //!
 //! impl RaftNetwork<MyConfig> for MyNetwork {
 //!     async fn install_snapshot(...) { ... }  // chunk-based RPC
@@ -34,7 +34,7 @@
 //! chunks. It derefs to the inner Raft, so all standard methods are available.
 //!
 //! ```ignore
-//! use openraft_network_v1::ChunkedRaft;
+//! use openraft_legacy::network_v1::ChunkedRaft;
 //!
 //! let inner = openraft::Raft::new(...).await?;
 //! let raft = ChunkedRaft::new(inner);
@@ -46,12 +46,12 @@
 //! raft.install_snapshot(req).await?;
 //! ```
 
-mod adapt_v1_to_v2;
+mod adapter;
 mod chunked_raft;
 mod network;
 mod sender;
 mod streaming;
 
-pub use adapt_v1_to_v2::Adapter;
+pub use adapter::Adapter;
 pub use chunked_raft::ChunkedRaft;
 pub use network::RaftNetwork;
