+++ legacy/src/network_v1/network.rs
@@ -28,7 +28,7 @@ use openraft_macros::add_async_trait;
 /// [`RaftNetworkV2`](openraft::network::v2::RaftNetworkV2) instead.
 ///
 /// To use this trait with openraft, wrap your implementation with
-/// [`Adapter`](crate::Adapter) to convert it to `RaftNetworkV2`.
+/// [`Adapter`](super::Adapter) to convert it to `RaftNetworkV2`.
 ///
 /// A single network instance is used to connect to a single target node. The network instance is
 /// constructed by the [`RaftNetworkFactory`](`openraft::network::RaftNetworkFactory`).
