+++ legacy/src/network_v1/streaming.rs
