+++ tests/tests/snapshot_streaming/t10_api_install_snapshot.rs
@@ -6,7 +6,7 @@ use openraft::Config;
 use openraft::Vote;
 use openraft::raft::InstallSnapshotRequest;
 use openraft::storage::SnapshotMeta;
-use openraft_network_v1::ChunkedRaft;
+use openraft_legacy::network_v1::ChunkedRaft;
 
 use crate::fixtures::RaftRouter;
 use crate::fixtures::log_id;
